formatting and change map to <String, Double> instead of <String, Integer> plus casts to ints

+++ src/BaseModel.java
@@ -1,7 +1,6 @@
 import java.util.Collection;
 import java.util.Map;
 
-
 /**
  * This the abstract model class that every type of model will extend
  * 
@@ -11,19 +10,22 @@ import java.util.Map;
 
 public abstract class BaseModel {
 	private int numPointsForNeighbor;
-	
-	public BaseModel(Map<String, Integer> parameters, int points) {
+
+	public BaseModel(Map<String, Double> parameters, int points) {
 		numPointsForNeighbor = points;
 	}
-	
-	// Takes in a cell and updates it's future state based on it's neighbors and the rules of the model
-	public abstract Cell updateFutureState(Cell cellToUpdate, Collection<Cell> neighbors);
-	
+
+	// Takes in a cell and updates it's future state based on it's neighbors and
+	// the rules of the model
+	public abstract Cell updateFutureState(Cell cellToUpdate,
+			Collection<Cell> neighbors);
+
 	public int getSharePointsForNeighbor() {
 		return numPointsForNeighbor;
 	}
-	
-	public int countNeighbors(Cell cellToUpdate, int state, Collection<Cell> neighbors) {
+
+	public int countNeighbors(Cell cellToUpdate, int state,
+			Collection<Cell> neighbors) {
 		int neighborsWithState = 0;
 		for (Cell c : neighbors) {
 			if (c.getCurrentState() == state) {
